pyproject.toml
[tool.poetry]
name = "boa_ai"
version = "0.1.0"
description = ""
authors = ["Jannis Straus <jannis.straus@uk-essen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
dicomweb-client = "^0.50.4"
fastapi = "^0.115.0"
pydicom = "^2.0.0"
celery = {extras = ["s3"], version = "^5.0.3"}
simpleitk = "^2.4.0"
pydicom-seg = "^0.4.1"
PyJWT = "^1.7.1"
pandas = "^1.5.2"
openpyxl = "^3.1.5"
body-organ-analysis = "0.1.3"
uvicorn = "^0.12.3"
torch = "2.1.1"
# body-organ-analysis = { path = "./body-organ-analysis" }
tritonclient = {extras = ["grpc", "http"], version = "2.44.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
